Fixed a bug where some shapes didn't have a color. Colors are now serialized to RGB values between 0 and 255 (instead of between 0 and 1).

+++ pdfact-cli/src/main/java/pdfact/cli/pipes/serialize/PdfJsonSerializer.java
@@ -678,7 +678,7 @@ public class PdfJsonSerializer implements PdfSerializer {
     JSONObject colorJson = new JSONObject();
     if (color != null) {
       String colorId = color.getId();
-      float[] rgb = color.getRGB();
+      int[] rgb = color.getRGB();
 
       if (colorId != null && rgb != null && rgb.length == 3) {
         colorJson.put(ID, colorId);
